+++ app/Http/Controllers/Admin/EgresadosAdminController.php
@@ -11,6 +11,7 @@ use App\Models\Configuracion;
 use App\Models\Egresado;
 use App\Repositories\Admin\InscripcionRepository;
 use Illuminate\Http\Request;
+use App\Models\Examen;
 
 use function PHPUnit\Framework\returnSelf;
 
@@ -124,6 +125,12 @@ class EgresadosAdminController extends BaseController
         public function destroy($id)
 {
     $inscripto = Egresado::findOrFail($id);
+
+    //verificar que no tenga mesas futuras
+    if(Examen::where('id_alumno',$inscripto->id_alumno)->where('fecha','>',date('Y-m-d'))->exists()) {
+        return redirect()->route('admin.inscriptos.index')
+            ->with('error', 'No se pudo eliminar la inscripción porque el alumno tiene mesas de examen futuras.');
+    }
     $inscripto->delete();
